Separate message logic by type
