+++ server/writer/index.ts
@@ -1,0 +1,11 @@
+import * as proposal from './proposal';
+import * as vote from './vote';
+import * as settings from './settings';
+import * as deleteProposal from './delete-proposal';
+
+export default {
+  proposal,
+  vote,
+  settings,
+  'delete-proposal': deleteProposal
+};
